refactor: 티켓 판매원에게 티켓 판매 책임을 주어 자율성 증진
