+++ app/src/main/kotlin/dev/dominiqn/objects/ticketing/Theater.kt
@@ -10,14 +10,6 @@ class Theater(
      * 관람객 입장
      */
     fun enter(audience: Audience) {
-        if (audience.bag.hasInvitation) {
-            val ticket = ticketSeller.ticketOffice.getTicket()
-            audience.bag.ticket = ticket
-        } else {
-            val ticket = ticketSeller.ticketOffice.getTicket()
-            audience.bag.minusAmount(ticket.fee)
-            ticketSeller.ticketOffice.plusAmount(ticket.fee)
-            audience.bag.ticket = ticket
-        }
+        ticketSeller.sellTo(audience)
     }
 }
